% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEP_Exports.R
\name{IEP_Chartbook}
\alias{IEP_Chartbook}
\title{Create Excel Chartbook}
\usage{
IEP_Chartbook(filepath)
}
\arguments{
\item{filepath}{Character. Path where the chartbook file should be saved.}
}
\value{
An Excel workbook object
}
\description{
Creates a new Excel workbook for chartbook purposes with a default worksheet.
}
