% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEP_EditProfile.R
\name{IEP_ShowProfile}
\alias{IEP_ShowProfile}
\title{Show Current R Profile Settings}
\usage{
IEP_ShowProfile(scope = c("user", "project"))
}
\arguments{
\item{scope}{Character. Either "user" or "project"}
}
\description{
Display the contents of the active .Rprofile without opening it
}
