% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchIEP-package.R
\docType{package}
\name{researchIEP}
\alias{researchIEP}
\alias{researchIEP-package}
\title{researchIEP: Research Tools for the Institute for Economics and Peace}
\description{
The researchIEP package provides standardized tools and functions
for IEP research projects, including project setup, data processing,
and visualization utilities.
}
\section{Main Functions}{

\itemize{
\item \code{\link{IEP_ProjectFolders}}: Create standardized project structure
\item \code{IEP_LibraryLoader}: Load multiple packages efficiently
\item \code{IEP_ChartTheme}: Apply IEP chart styling
}
}

\section{Getting Started}{

To create a new project with IEP structure:
\preformatted{
library(researchIEP)
IEP_ProjectFolders("path/to/project", type = "simple")
}
}

\references{
\url{https://www.economicsandpeace.org}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tmorgan-IEP/researchIEP}
  \item Report bugs at \url{https://github.com/tmorgan-IEP/researchIEP/issues}
}

}
\author{
Institute for Economics and Peace
}
\keyword{internal}
